requirements.txt
streamlit>=1.28.0
langchain>=0.0.300
langchain-openai>=0.0.2
pypdf2>=3.0.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
reportlab>=4.0.0
tiktoken>=0.5.0
